+++ design-system/popper/src/__tests__/unit/index.tsx
@@ -5,7 +5,7 @@ import { mount } from 'enzyme';
 import { Popper as PopperCompo } from '../../index';
 
 jest.mock('popper.js', () => {
-  const PopperJS = require.requireActual('popper.js');
+  const PopperJS = jest.requireActual('popper.js');
 
   return class Popper {
     static placements = PopperJS.placements;