+++ editor/editor-core/src/__tests__/integration/media/_utils.ts
@@ -4,7 +4,7 @@ export async function waitForNumImages(page: Page, n: number) {
   await page.waitUntil(async () => {
     const images = await page.$$('.ProseMirror [data-testid="media-image"]');
     return images.length >= n;
-  });
+  }, 'waitForNumImages failed');
 
   return await page.$$('.ProseMirror [data-testid="media-image"]');
 }
@@ -15,7 +15,7 @@ export async function waitForAtLeastNumFileCards(page: Page, n: number) {
       '.ProseMirror [data-testid="media-file-card-view"][data-test-status="complete"]',
     );
     return fileCards.length >= n;
-  });
+  }, 'waitForAtLeastNumFileCards failed');
 
   return await page.$$(
     '.ProseMirror [data-testid="media-file-card-view"][data-test-status="complete"]',