+++ editor/editor-core/src/__tests__/unit/ui/PortalProvider.tsx
@@ -1,5 +1,5 @@
 const mockReactDom = {
-  ...jest.requireActual('react-dom'),
+  ...jest.requireActual<Object>('react-dom'),
   unmountComponentAtNode: () => {},
 };
 jest.mock('react-dom', () => mockReactDom);