+++ editor/renderer/src/__tests__/unit/index.ts
@@ -13,7 +13,7 @@ import { defaultSchema as schema } from '@atlaskit/adf-schema';
  */
 jest.mock('@atlaskit/editor-common/validator', () => ({
   __esModule: true,
-  ...jest.requireActual('@atlaskit/editor-common/validator'),
+  ...jest.requireActual<Object>('@atlaskit/editor-common/validator'),
 }));
 import * as common from '@atlaskit/editor-common/validator';
 import { renderDocument, Serializer } from '../../index';